rutabaga_gfx: remove unnecessary casts
Fixes clippy warnings in Rust 1.68.

BUG=b:276487055
TEST=tools/clippy # with rust 1.68

Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4391109
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>

+++ rutabaga_gfx/src/gfxstream.rs
@@ -343,7 +343,7 @@ impl Gfxstream {
 
     fn export_blob(&self, resource_id: u32) -> RutabagaResult<Arc<RutabagaHandle>> {
         let mut stream_handle: stream_renderer_handle = Default::default();
-        let ret = unsafe { stream_renderer_export_blob(resource_id as u32, &mut stream_handle) };
+        let ret = unsafe { stream_renderer_export_blob(resource_id, &mut stream_handle) };
         ret_to_res(ret)?;
 
         // Safe because the handle was just returned by a successful gfxstream call so it must be
@@ -534,7 +534,7 @@ impl RutabagaComponent for Gfxstream {
         let (iovecs, num_iovecs) = match buf {
             Some(buf) => {
                 iov.base = buf.as_ptr() as *mut c_void;
-                iov.len = buf.size() as usize;
+                iov.len = buf.size();
                 (&mut iov as *mut RutabagaIovec as *mut iovec, 1)
             }
             None => (null_mut(), 0),
